Use 16 bit keyspace

+++ hydra/keyspace.h
@@ -6,7 +6,7 @@
 namespace hydra {
 
 struct keyspace_t {
-  using value_type = uint8_t;
+  using value_type = uint16_t;
 
   static_assert(std::numeric_limits<value_type>::digits,
                 "You have to implement std::numeric_limits<keyspace_t>.");
